Fix a bug of MissingMethodException

+++ src/main/groovy/com/scorsi/gameengine/entities/Visible.groovy
@@ -17,9 +17,9 @@ trait Visible implements Positionable, Sizable {
             g.drawImage(image,
                     (position.x - handler.camera.offset.x) as Integer,
                     (position.y - handler.camera.offset.y) as Integer,
-                    width, height, null)
+                    width as Integer, height as Integer, null)
         else
-            g.drawImage(image, position.x as Integer, position.y as Integer, width, height, null)
+            g.drawImage(image, position.x as Integer, position.y as Integer, width as Integer, height as Integer, null)
     }
 
 }
